Use git rev-parse over ls-remote
This commit is slightly messy. Added "unwrap" flag to run_command which
only applies when capture_output = true. Function docstring not updated.
ls-remote only contained some commit hashes, but rev-parse does more or
less exactly what we want.

stderr is printed in the output, but this is OK. This commit gave me the
idea that it would be nice if errors could be collected, but it would
seem like quite the effort.

When debug mode was set, every log message was confusingly using the
DEBUG prefix. Reverted this behavior so that it behaves more as you
would expect.