Fix clone and caching
Still just a bit messy, but it is due to something that can't easily be
changed. git rev-parse requiring a full git clone is unhelpful. Working
toward a single solution to resolve the refspec would be nice to have,
but the lua function is nicer to use. The shell snippet hurts the build
because it means the script is no longer cross platform, not to
mention the error handling is more difficult.

Returning a function that should be evaluated after the clone has
completed would be one way to address this issue. Does feel like
spaghetti, but it would solve the problem.