Add finish_processing function
Added a function to organize the two processing steps that needed to
happen after the clone, and also de-duplicated some work and removing
the requirement on shell.

Overall I wonder whether the error handling, and maybe even the naming
of some functions could be tweaked, but the project feels much better
now in terms of where it is.

I believe it should be stable from here for at least a little while,
even if the docstrings may need to be updated.